+++ src/client/routes/devices/components/widgets/Logs.js
@@ -3,7 +3,7 @@ import moment from 'moment'
 import { List } from 'immutable'
 import { connect } from 'react-redux'
 
-import { cleanLogs } from '/routes/devices/modules/actions'
+import { cleanLogs } from '/routes/devices/actions'
 
 class Logs extends Component {
 	shouldComponentUpdate (nextProps) {